updated with logging and exception handling

+++ SpringBootRest/src/main/java/com/webonise/springBoot/controller/StudentController.java
@@ -2,6 +2,8 @@ package com.webonise.springBoot.controller;
 
 import java.util.List;
 import java.util.Optional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,8 +12,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.webonise.springBoot.dao.StudentRepository;
+import com.webonise.springBoot.exception.StudentNotFoundException;
 import com.webonise.springBoot.model.Student;
 
 @RestController
@@ -19,6 +21,7 @@ public class StudentController {
 
 	@Autowired
 	StudentRepository studentRepository;
+	Logger log = (Logger) LoggerFactory.getLogger(StudentController.class);
 
 	@RequestMapping(path = "/students")
 	public List<Student> home() {
@@ -27,25 +30,32 @@ public class StudentController {
 
 	@RequestMapping("/student/{studentId}")
 	public Optional<Student> getStudent(@PathVariable("studentId") int studentId) {
+		if (!studentRepository.existsById(studentId))
+			throw new StudentNotFoundException();
 		return studentRepository.findById(studentId);
 	}
 
 	@PostMapping(path = "/student")
-	public Student addStudent(@RequestBody Student student) {
+	public Boolean addStudent(@RequestBody Student student) {
 		studentRepository.save(student);
-		return student;
+		log.info("New student is added :" + student.toString());
+		return true;
 	}
 
 	@DeleteMapping(path = "/student/{studentId}")
-	public String deleteStudent(@PathVariable int studentId) {
+	public Boolean deleteStudent(@PathVariable int studentId) {
+		if (!studentRepository.existsById(studentId))
+			throw new StudentNotFoundException();
 		Student student = studentRepository.getOne(studentId);
 		studentRepository.delete(student);
-		return "deleted";
+		log.info("student is removed from database :" + student.toString());
+		return true;
 	}
 
 	@PutMapping(path = "/student")
-	public Student SaveOrUpdateStudent(@RequestBody Student student) {
+	public Boolean SaveOrUpdateStudent(@RequestBody Student student) {
 		studentRepository.save(student);
-		return student;
+		log.info("Student is updated :" + student.toString());
+		return true;
 	}
 }
